refactor: Rename and reorganize GitHub-related modules and update imports
